vchan library support for new blocking semantics
The new blocking semantics for the vm side of the vchan connection rely on the direct values of the amount of data ready to read and the amount of buffer space left to write.

The vchan status codes have thus been removed (as they are no longer needed), and the vchan_alert_t stores the connection data, instead of a status code.