sel4arm-vmm: Fix iospace metadata corruption bug
The shadow data for iospaces was getting allocated using realloc, this
meant that the address of the shadow data could be changed in subsequent
calls to realloc. This causes the reference that the generic vspace
implementation keeps to the data to become invalid. We fix this by doing
a seperate allocation for each iospace, and only use realloc to grow the
list of pointers to this shadow data