sel4arm-vmm: Removed VPCI and IOPORT config option
Removed the use of the CONFIG_LIB_SEL4_ARM_VMM_VPCI_SUPPORT and
CONFIG_LIB_SEL4_ARM_VMM_IOPORTS_SUPPORT config options. Instead
we leave the pci and io_port fields in the vm struct as
uninitialised pointers. This reduces the size of the vm structure
if the refs aren't used. Additionally updated usages of the
pci and io_ports field to reflect change.