libsel4vm: Added libsel4vm
This is intended to be a vm library, unifying the codebases of
libsel4arm-vmm and libsel4vmm (x86) into a single project. This
an initial commit that bases all the code from the libsel4vmm and
libsel4arm-vmm into a single library, where further work will
focus on refactoring and merging the source code into a unified
interface.