libsel4vm: VM run interface
Refactored the run invocation into 'vm_run'. This updates the x86
and arm source to implement their runtime event loops when the
'vm_run' interface is used. The intented functionality of vm run
is such that it will start the vm catching vm exits. This will
either handle the exits internally or forward it to the user.