libsel4vm: Refactored arm vm event handler
Refactored the arm vm event/exit handler. This refactors the vm
exit handler code into a series of individual handlers for the
various vm exit cases. The overall intention with these changes
is to make the vm exit handler more modular and to reduce the
complexity of the vm_run function.