libsel4vm: Enabled vm_run to exit to caller
Introduced vm exit constants that can be used to return to the
library caller (vmm) for handling specific vm exits. Updated the
exit handling functions and vm run loop to handle cases where it
may exit and return an exit reason to the vmm. This is an initial
change that will allow the vm to reduce policy and refactor out
various drivers from the library such that it can be handled and
managed by the vmm.