sel4vmmplatsupport: Moved plat devices from sel4vm
Added an initial library 'sel4vmmplatsupport' to contain platform
specific drivers and utilities. This is introduced now to hold
the various arm platform devices. This library DOES NOT currently
compile. The code moved needs to be re-evaluated and updated to
to use a new memory interface and device registration interface.