sel4vmmplatsupport: Moved unsupported device utils
Moved unsupported device utils into sel4vmmplatsupport. This code
will NOT compile. The code moved needs to be re-evaluated and
updated to use a new memory interface and device registration
interface. This is being moved now as its currently incompatible
with the new memory interface and will need to be re-assessed in
future updates.