libsel4vmmcore: Removed IOPort passthrough type
Removed the ability to add passthrough ioports. This functionality
isn't utilized in the context of the libsel4vmmcore library.
Neither is the passthrough function made publicly available.
Adding and configuring passthrough ioports is architecture/platform
specific and is left to external libraries and applications to
establish. This keeps the sel4vmmcore ioports utility a generic
management datastructure.