libsel4vmmcore: Configure virtio_net BAR emulation
Depending on how accesses are managed to the PCI Config Address
space, provide the option to configure how read and writes are
performed on the virtio net BARs. The new configuration option,
when enabled, allows read and writes to the BARs to be emulated,
proxying BAR accesses to the virtio net pci device. This
is needed as ARM and x86 manage and emulate accesses to the PCI
Config Address space differently.