libsel4vm: Moved device drivers to vmmplatsupport
Moved the arm device drivers from libsel4vm out to
libsel4vmmplatsupport.These are optional devices a VMM may choose
to initialise and hence do not need to be managed by libsel4vm.
Note these devices do NOT compile and need to be updated to use
the new memory interface for reservation management and
fault handling. This is an intermediate commit in preperation for
updating the device drivers.