libsel4vm: Consolidated VCPU fault interface
Added a new utility interface to extract general fault information
out of a vcpu object. This abstracts the architecture specific
interfaces for retrieving fault information. This interface is
introduced such that the architecture specific interfaces can
be made private and to also create a utility that future drivers
can use across different platforms.