libsel4vm: Remove arch_data in mem fault callback
In the mem fault callback interface we replace the passing of the
guest_memory_arch_data interface with the faulting vcpu object.
Since we now have a vcpu fault interface we no longer need to
construct a custom datastructure to pass vcpu fault information.
The callback can rather query the needed information through the
vcpu fault interface. Updated the sources to reflect the interface
change.