libsel4vm: Updated 'create_emulated_reservation'
Updated the name and parameters of the
'create_emulated_reservation_frame' utility. This is now
'create_allocated_reservation_frame' and additionally allows the
user to pass in the rights of the vm vspace mapping. This
generalises the function, which whilst it can still support the
creation of emulated frames, it can be used to create other types
of shared vm-vmm vspace mappings.