libsel4vmmplatsupport: arm devices use libsel4vm
Updated the arm device drivers to use libsel4vm for
registering their device memory regions and their fault handlers.
This is a general update to the arm architecture devices to use
the updated interfaces in libsel4vm and libsel4vmmplatsupport.
Prior to this update the devices would not compile/work.