libsel4vm: Removed usage of x86 arch.guest_image_t
Removed the usage of the vm.arch.guest_image.t datastructure.
This has be reduced to a structure for managing the guest boot info
(which is intended to be temporary and further refactored). The
other fields of guest_image have been parameterised. This stages
way for refactoring and moving out the image loading code into
libsel4vmmplatsupport.