libsel4vm: Moved guest boot src to vmmplatsupport
The boot_guest interface and its accompanying components (e820,
bootinfo, acpi) are platform specific interface. This should rather
be code a VMM can opt in to using for initialising its x86 VM boot
environment. This commit simply moves the source and does not
compile. Future work will update the source to a working state such
that it uses the libsel4vm api.