libsel4vm: Moved VM tcb object into vcpu object
There is a strong association between a VCPU and the TCB it is
binded to. Thus the TCB object should exist in the context of the
VCPU it is associated with. In addition this makes the VM library
open to possible SMP extensions. This commit updates the
initialisation of the TCB object when creating a VCPU and
the means of accessing TCB objects.