libsel4vm: Refactored and moved arm reboot hooks
Moved out the ARM reboot hooks interface into sel4vmmplatsupport.
This is rather a utility a VMM should use if it wishes to create
reboot callbacks around a VM it manages. This commit larges
refactors the code into an independent interface in
sel4vmmplatsupport and updates usages of reboot hooks.