libsel4vmmplatsupport: Trap A57 CPUACTLR_EL1
Adding a handler to the aarch64 sysreg fault infrastructure
concerned with the case when an ARM Cortex A57 core attempts to
access the CPU Auxiliary Control Register from EL1. Some platforms
(e.g tx2) choose to write to the register on powerup to set
implementation defined CPU configuration and control options. We
choose to simply ignore the write and proceed with execution.