libsel4vm: irq controller interface takes vcpu's
Updated the guest irq controller interface to use vcpu handles
over vm handles when it comes to irq management (inject and ack).
Using the vcpu handle is needed when injecting interrupts that
are designated to specific cores e.g. PPI's. The vm handle can
additionally be derived from the vcpu if it is needed in the
interrupt controller driver.