libsel4vm: Program VCPU Affinity on creation
We set the affinity of our created vcpu TCB's to correspond with
the core it represents. In addition we modify the BOOT_VCPU's
Virtualization Multiprocessor ID Register to indicate we are using
ARMv7 Multiprocessing Extension functionality and the performance
of our logical cores at the lowest affinity level is
interdependent.

Co-authored-by: Anna Lyons <anna@gh.st>