libsel4vm: Revised arm fault types
Added a new field in the arm fault data-structure that
consolidates the different types the fault can represent. This
being a DATA,PREFETCH or VCPU fault. It previously only assumed
the fault could be a data or prefetch fault when advancing or
manipulating the faults state. Updated various function names to
reflect the new types.