libsel4vm: arm context interface uses fault state
Updated the guest arm context interface to account for the
fault interface if the vcpu is in a middle of a fault. Prior
this would directly write to the TCB and potentially de-sync
with the faults state (whom would eventually override anything
written). If the vcpu is in the middle of a fault we now update
and read its cached state as this will eventually be syned to the
TCB when the vcpu resumes.