libsel4vm: Created 'vcpu_start' interface
This commit introduces a new interface function, 'vcpu_start',
that can be invoked to configure and start a newly created vcpu
object. This change refactors over the private 'vcpu_resume'
interface. The key impact over this change is the user will now
need to invoke 'vcpu_start' for each vcpu rather then relying
on 'vm_run' to do so. This adds the needed flexibility for
operations such a deferring the starting of a vcpu e.g. until
necessary psci/pm operations are performed.