libsel4vm: arm vgic lr overflow for each vcpu
Updated the vgic management datastructure to have an LR overflow
list for each VCPU in the system. This being needed to maintain the
IRQs we inject on each core for SMP-based configurations.
Additionally added helper functions to queue and dequeue the
overflowed irqs to simplify to implementation of now managing
multiple LR overflow lists.

Co-authored-by: Anna Lyons <anna@gh.st>