libsel4vm: Restructure the GICv2 Distributor Map
Since we now are supporting ARM SMP VM's, certain registers in the
GICv2 distributor need to be banked for each vcpu. This change
restructures the layout of the registers to account for the
registers that need to be banked and updates the usages of
accessing these registers to access the correct banked register
for a given vcpu context.