libsel4vm: Completely emulate vgicv2 distributor
Due to refactoring the vgicv2 distributor register map datastructure
to account for register banking, its memory region can no longer
consistently map to the physical memory layout of the gic distributor.
This meaning we can no longer pass read rights to the VM (as we
need to emulate all accesses). This change removes the mapping
of a read-only emulated frame and now reserves the distributor
region with no access rights. This is a breaking change, with
subsequent changes coming to overhaul the distributors memory fault
handling.