libsel4vm: Overhaul GIC distributor fault handling
This change refactors the fault handling of the gic distributor
to account for its new layout and the need to now emulate read
accesses. As accesses to the GIC distributor now need to account
for banked registers, using a common set of 'ACTION's to access the
new map datastructure is no longer practical. The memory fault
handler now decodes and emulates accesses on a per-register basis.
This in addition gives finer grain control over emulation of
specific distributor registers to take into account the vcpu
performing the access.