libsel4vmmplatsupport: Dynamic ioport allocation
Updated the ioports interface to support semi-dynamic allocation
of ioport ranges. The interface has been extended to take a
starting address by which the interface can allocate free ioport
ranges. For the time being this is a simple bump allocator with
no ability to free up allocated ioport ranges. The user can specify
whether they want to add an ioport for a specific range or let
it allocate at a free range.