libsel4vmmplatsupport: Allocate ioports objects
Updated the ioport interface such that registered ioport ranges
are allocated as objects that can be returned to the
user. This is useful if the interface user chooses to register
an ioport handler for any free unallocated range but wishes to
later know what range was reserved i.e. to encode a device into a
PCI config space.