libsel4vmmplatsupport: Updated vpci irq map
Updated the generation of the vpci irq map to take into account
the address cells value defined in the fdt interrupt controller
node. This is an intermediate solution to resolving the number
cells to insert into the parent interrupt specifier. Ideally we
want to extract the '#address-cells' value out of the fdt irq
controller node (or default to 1 if not defined).