libsel4vmmplatsupport: Fix an off-by-one error
Previously, the 'end' of an IO port range was not inclusive. This would
result in a bug where one IO port whose value is the 'end' of a range
would be mistaken as being part of that range when it should not be.
This commit fixes it so that the 'end' of an IO port range is now
inclusive. Also harden the range overlap check.