cross_vm_connection: Round up event BAR size
As our vpci device doesn't support emulating remapping of device
resource physical addresses by the guest, we need to create our devices
in a way that Linux won't decide to remap them. This means making each
BAR the same size.  For Cross-VM connectors we have a dataport and event
resource. We describe both as being the size of the dataport, but only
create a single page for the event port.  This assumes that the guest
will only access the first word of the event port for signalling events
and we have a large enough address range to spend on making the event
resource as large as the dataport one.