cross_vm_connectors: Add a device name register
The event PCI BAR allocates a 4K frame but only uses the first 8 bytes
for sending and receiving events. The next 50 bytes are now used to
store a device name that a guest OS can use to identify different
devices.

A default of "connector" is used because this was the name that the
Linux kernel module would give to all devices before this change.