Ensure correct bounds of VM exit handlers
We need to ensure that we don't search for a VM exit reason outside the
expected set of reasons when we access or call the handlers.

This ensures the array has the correct length even if we don't handle
reasons with higher numeric identifiers and ensures that we don't access
the array outside of its bounds.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>