Ensure write back of vm state
The gcc we currently use appears not to correctly write back the state
of a VM or the VM exit conditions when fully inlined and optimised.

We add a few judicious uses of `volatile` to ensure that when the
code suggests values should be written back to the state, they do
actually get written back, even for the most aggressively inlined code.

There is clearly something wrong with GCC here as it appears to be
inlining the entire vm loop and losing the fact that the VMEnter and
Wait syscalls can modify memory.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>