HYPR-161 Enlarge PCI memory window
Since the virtio PoC exposes all of the user-VM's memory to the
driver-VM via a CAmkES dataport, the PCI memory window must be
large enough to accommodate the guest memory. Right now we have
128 MB configured for the user-VM, but to be on the safe side,
lets use 512 MB here.

The address 0x50000000 is picked because the driver-VM currently
occupies 0x40000000 - 0x47ffffff and the user-VM occupies
0x48000000 - 0x4fffffff. In other words, nothing is there
currently. One would think this address is specific to the
vspace of the VM and hence it is enough if the current VM does
not have any other kind of memory -- but because the PCI memory
window is statically configured in the header and is the same for
every VM, we need to compute a union of all used address ranges
and then select some range outside of this union. Perhaps some
day we could have per-VM defined PCI memory window?

It appears that the PCI memory window shall be within the 32-bit
address space until we introduce 64-bit addressing to seL4's
virtual PCI support.

Signed-off-by: Hannu Lyytinen <hannu.lyytinen@unikie.com>