TEMPORARY HACK HYPR-161 Use 2 MB large pages
It appears CAmkES machinery uses the largest page size it can. In other
words, when we declare a 128 MB dataport, it creates 2 MB large pages
for that. The problem is the code expects 4 kB pages.

We can make the 128 MB dataport work if we just hardcode the page
size, but sooner or later we need to support multiple page sizes.

Signed-off-by: Hannu Lyytinen <hannu.lyytinen@unikie.com>