libsel4vmmplatsupport: fix PCI passthrough on ARM
BAR size resolution was emulated in two places, the fault handler and
the backend created with vmm_pci_create_passthrough_bar_emulation().
Move to vmm_pci_create_bar_emulation() on ARM VM as well, and remove
the former. This way there is no need for BAR hack in the fault handler
and the backend created with vmm_pci_create_passthrough() works now on
ARM.

Also fixes the write size of the PCI fault handler, as it was discarded
previously and fixed size of 32 bits was always used.

As a side effect, the cookies for FDT generation are not necessarily
the same type anymore. Therefore we cannot consult them for interrupt
pins and lines, but on ARM the supported devices are virtio ones only
and currently they all use the same pin and the same line. Let the VM
specify those as an intermediate solution.

Signed-off-by: Hannu Lyytinen <hannu.lyytinen@unikie.com>