libsel4vmmplatsupport: remove deprecated emulation
Having a PCI emulation that supplies the configuration header from an
external data structure, except the BAR registers, seems to be a
leftover from the era when x86 and ARM VMs lived in separate trees.
This emulation was necessary to avoid doing double emulation on BAR
registers, since the fault handler did that too. The latter is now fixed
and this emulation serves no purpose anymore.

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>