libsel4vm: refactor vGIC for readability
This is mostly rearranging the code so that the IRQ queueing mechanism
becomes more apparent, making also the maintenance handler much simpler.

The ring buffer is implemented in a bit more readable way and it is
decoupled from the rest; this makes it easier to support IRQ priorities
in the future.

Signed-off-by: Hannu Lyytinen <hannux@ssrc.tii.ae>