libsel4vm: add vuart for zynqmp
This implementation has made several design decisions:
    * No concept of an "active" uart
    * The vuart TEMPTY interrupt status flag is kept asserted
    * There is a predefined console
    * Default serial port is UART0

The TEMPTY interrupt status flag is always asserted because it is set
when writing the TX FIFO, but cleared when the interrupt is
acknowledged. This causes a problem if a write to the TX FIFO occurs in
the guest's interrupt handler, because the interrupt could be
acknowledged after writing to the TX FIFO.

The vuart device immediately flushes all writes to the TX FIFO, so it
is always empty. This commit will keep the TEMPTY flag set
indefinitely to more closely emulate real hardware behavior.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>