vmmplatsupport, vesa: fix vesa mapping
The previous behavior would only attempt to map the vesa framebuffer and
set the screen info struct if the vbe protection mode interface was
successfully mapped in. This removes the dependency on the vbe protected
mode interface. It is not clear that it is needed as vesa framebuffer
passthrough works without it.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>